add functional invocation path
